finding temp by id
