+++ trash.py
@@ -1,0 +1,49 @@
+#json_data = json.load(f)
+#json_data=json.load(f)
+#print(json_data)
+#d=str(json_data)
+#print(json_data[0])
+#print(type(d))
+#e = open("myfile.txt", "w", encoding='utf-8')
+#e.write(d)
+#e.close()
+
+#print(len(json_data))
+#i=0
+#list=[]
+#while i <len(json_data)-1:
+#    i += 1
+#    #print (i)
+#   if type(json_data[i]['id']) == int:
+#       list.append(json_data[i]['id'])
+#   else:
+#       break
+
+#print(len(get_weather(i)))
+#for n in list:
+#    print(n)
+
+#print(type(list))
+#i=820
+#while i<835:
+#    print(i)
+#    i+=1
+#    if type(get_weather(i)) == str:
+#       list.append(get_weather(i))
+#       print(get_weather(i))
+        #else:
+        #    break
+#list.append(get_weather(833))
+#print(list)
+
+#f = open('city.list.json',"r", encoding="utf8")
+
+#print(json_dumps)
+#input("type anything to end")
+
+#f = open("city.list.json", "r")
+#print(f.read())
+#i=1
+#while i<1000:
+#    try:
+#        resp = requests.get(url, params=parameters)
